#347: Rename Profile Page (and associated pages)
Rename the profile page to UserProfile.tsx to clarify between a "user profile" and a "social profile."
